refactor: reservation 데이터 프로퍼티 변경

+++ app/src/main/java/woowacourse/movie/presentation/ui/reservation/ReservationActivity.kt
@@ -3,8 +3,8 @@ package woowacourse.movie.presentation.ui.reservation
 import android.content.Context
 import android.content.Intent
 import woowacourse.movie.R
+import woowacourse.movie.data.repository.ReservationRepositoryImpl
 import woowacourse.movie.databinding.ActivityReservationBinding
-import woowacourse.movie.domain.dummy.DummyReservation
 import woowacourse.movie.domain.dummy.DummyTheater
 import woowacourse.movie.domain.model.Reservation
 import woowacourse.movie.presentation.base.BaseMvpBindingActivity
@@ -18,7 +18,7 @@ class ReservationActivity : BaseMvpBindingActivity<ActivityReservationBinding>()
     }
 
     override fun initStartView() {
-        val id = intent.getIntExtra(PUT_EXTRA_KEY_RESERVATION_ID, DEFAULT_RESERVATION_ID)
+        val id = intent.getLongExtra(PUT_EXTRA_KEY_RESERVATION_ID, DEFAULT_RESERVATION_ID)
         presenter.loadReservation(id)
     }
 
@@ -34,11 +34,11 @@ class ReservationActivity : BaseMvpBindingActivity<ActivityReservationBinding>()
 
     companion object {
         private const val PUT_EXTRA_KEY_RESERVATION_ID = "reservationId"
-        private const val DEFAULT_RESERVATION_ID = -1
+        private const val DEFAULT_RESERVATION_ID = -1L
 
         fun startActivity(
             context: Context,
-            reservationId: Int,
+            reservationId: Long,
         ) {
             val intent = Intent(context, ReservationActivity::class.java)
             intent.putExtra(PUT_EXTRA_KEY_RESERVATION_ID, reservationId)
