+++ app/src/main/java/woowacourse/movie/presentation/ui/seatselection/SeatSelectionContract.kt
@@ -23,7 +23,7 @@ interface SeatSelectionContract {
 
         fun showTotalPrice(totalPrice: Int)
 
-        fun navigateToReservation(id: Int)
+        fun navigateToReservation(id: Long)
 
         fun showReservationDialog()
